webapp/requirements.txt
aiofiles==0.6.0
click==7.1.2
fastapi==0.63.0
h11==0.12.0
httptools==0.1.1
icalendar==4.0.7
imgkit==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
Pillow==8.2.0
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.16.0
pytz==2021.1
PyYAML==5.4.1
recurring-ical-events==0.1.21b0
six==1.15.0
starlette==0.13.6
typing-extensions==3.7.4.3
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
